requirements.txt
flask
numpy
pandas
jsonlib
beautifulsoup4
requests2